Update obs_test.table to add missing entries
ups/obs_test.table was missing some entries for packages
needed by unit tests and bin/genInputRegistry.py.
Note that there are additional entries, including ip_isr, pyfits and meas_algorithms,
that are required by code in data/utils. That code is only used to generate the
data repository in this package, and so is essentially never run. Adding
the extra entries to the ups table file would create needless extra dependencies,
and in the case of meas_algorithms it would create a circular dependency.