Improve TestMapper based on new daf_butlerUtils CameraMapper
Eliminated TestMapper.bypass_camera and all other x_camera
methods (most of which were commented out) because CameraMapper
now has better support for constructing cameras in subclasses.

Add a test to tests/map.py that the butler can get a camera.

As a result of the changes in CameraMapperm cameras are no longer
standardizable, so change the expected result in tests/map.py